requirements.txt
deap==1.4.1
graphdb-python==0.0.6
matplotlib==3.7.3
mysqlclient==2.2.0
numpy==1.23.5
pandas==1.5.2
PyQt5==5.15.10
python-dotenv==1.0.0
scikit-learn==1.3.2
seaborn==0.13.0
